Correctly set strides for old format.
For cropped field of views and old-format dcimgs, the strides are
incorrect: they should be set using the "bytes_per_row" field.  This PR
fixes this issue.

As an additional note, it appears that the images returned by dcimg are
transposed by 90Â, assuming that the notion of "row" should indeed be
horizontal.  (This shows up as the "weird" stride values.)  This could
be fixed in a later PR.